fix: Forbid deletion of current version in version expiration code
Log it instead.

+++ lib/Versions/GroupVersionsExpireManager.php
@@ -18,6 +18,7 @@ use OCP\AppFramework\Utility\ITimeFactory;
 use OCP\EventDispatcher\IEventDispatcher;
 use OCP\Files\FileInfo;
 use OCP\IUser;
+use Psr\Log\LoggerInterface;
 
 /**
  * @psalm-import-type InternalFolderOut from FolderManager
@@ -29,6 +30,7 @@ class GroupVersionsExpireManager {
 		private VersionsBackend $versionsBackend,
 		private ITimeFactory $timeFactory,
 		private IEventDispatcher $dispatcher,
+		private LoggerInterface $logger,
 	) {
 	}
 
@@ -68,6 +70,18 @@ class GroupVersionsExpireManager {
 				$versions = $this->versionsBackend->getVersionsForFile($dummyUser, $file);
 				$expireVersions = $this->expireManager->getExpiredVersion($versions, $this->timeFactory->getTime(), false);
 				foreach ($expireVersions as $version) {
+					if ($version->isCurrentVersion()) {
+						$this->logger->error(
+							'Current version of a groupfolders file was listed for deletion. Skipping.',
+							[
+								'folderid' => $version->getFolderId(),
+								'timestamp' => $version->getTimestamp(),
+								'mtime' => $version->getSourceFile()->getMtime(),
+								'sourcefilename' => $version->getSourceFileName(),
+							]
+						);
+						continue;
+					}
 					/** @var GroupVersion $version */
 					$this->dispatcher->dispatchTyped(new GroupVersionsExpireDeleteVersionEvent($version));
 					$view->unlink('/' . $fileId . '/' . $version->getVersionFile()->getName());
